(PieGraph Library Edit) In PieGraphView, grouped Integer, Double, Float logic together, but separate from Long logic. To prevent overflows, since piegraph display is all relative, average value is used to determine total and percentages. (Even if all values are MAX_VALUE for their class, if it's divided by size of the list before being added to total, it will never overflow)

+++ app/src/main/java/shuvalov/nikita/graphviewsexample/MyPieGraphables.java
@@ -6,17 +6,17 @@ import shuvalov.nikita.piechartview.PieGraphable;
  * Created by NikitaShuvalov on 3/1/18.
  */
 
-public class MyPieGraphables implements PieGraphable<Float> {
-    private Float someValue;
+public class MyPieGraphables implements PieGraphable<Long> {
+    private Long someValue;
     private String someName;
 
-    public MyPieGraphables(Float someValue, String someName) {
+    public MyPieGraphables(Long someValue, String someName) {
         this.someValue = someValue;
         this.someName = someName;
     }
 
     @Override
-    public Float getValue() {
+    public Long getValue() {
         return someValue;
     }
 }
